% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flags.R
\name{flags}
\alias{flags}
\alias{mklCxxFlags}
\alias{mklLdFlags}
\alias{onemklLibFlags}
\alias{onemklIncFlags}
\title{Compilation flags for oneMKL}
\usage{
mklCxxFlags()

mklLdFlags()

onemklLibFlags()

onemklIncFlags()
}
\description{
Output the compiler or linker flags required to build against `oneMKL`.
These functions are typically called from `Makevars` as follows:
```
PKG_CXXFLAGS += $(shell "${R_HOME}/bin/Rscript" -e "oneMKL::onemklIncFlags()")
PKG_LIBS += $(shell "${R_HOME}/bin/Rscript" -e "oneMKL::onemklLibFlags()")
```
\R packages using `oneMKL` should also add the following to their `NAMESPACE` file:
```
importFrom(oneMKL, onemklIncFlags)
importFrom(oneMKL, onemklLibFlags)
```
This is necessary to ensure that \pkg{oneMKL} is loaded and available.
}
