% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mklLibraryPath}
\alias{mklLibraryPath}
\title{Get the Path to Intel MKL Library}
\usage{
mklLibraryPath(name = NULL)
}
\arguments{
\item{name}{The name of Intel MKL library to be resolved.
It will be one of ``, ``, `` or ``.
When name is `NULL`, this function returns the path which contains Intel MKL.}
}
\description{
Get the Path to Intel MKL Library
}
