% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onemkl-package.R
\docType{package}
\name{onemkl-package}
\alias{onemkl-package}
\title{`oneMKL` Package}
\description{
The `oneMKL` package aims to provide header files and dynamic library files to R for usage.
We imported files from packages on Anaconda,  `mkl`, `mkl-include` and `intel-openmp`.
`oneMKL` supports windows and Linux only because the availablity of Anaconda `mkl` package.
}
