% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onemkl-package.R
\docType{package}
\name{onemkl-package}
\alias{onemkl-package}
\title{`oneMKL` Package}
\description{
The `oneMKL` package aims to empower R with Intel MKL like Microsoft R Open (MRO)
by leveraging the Anaconda packages, `mkl`, `mkl-include` and `intel-openmp`.
`oneMKL` aims to act like a plug-in for R users and support dynamic linkage for Rcpp users.
`oneMKL` supports windows and Linux only because the availablity of Anaconda `mkl` package.
Besides, `oneMKL` provides the ability to use Intel MKL in `Rcpp`,
so you can leverage MKL functions widely in `mkl_blas.h`, `mkl_cblas.h`, `mkl_spblas.h`,
`mkl_lapack.h` and `mkl_lapacke.h`.
}
